Fix build system and AI assistant stability
This commit addresses several issues found in the build system, test framework, and the AI assistant userspace program.

**Build and Test System Fixes:**

- **tests/Makefile**: Added the `-m32` flag to `CFLAGS_BASE` to ensure test binaries are compiled as 32-bit executables, matching the target OS architecture. This fixes a `futex` crash when running the 64-bit compiled tests.
- **tests/Makefile**: Corrected the `clean` target to safely remove test-related build artifacts without errors.
- **tests/scripts/run_all_tests.sh**:
    - Removed hardcoded absolute paths (`/workspace/ai-os`) and replaced them with relative paths, allowing the script to run in any directory.
    - Refactored the script to execute pre-compiled test binaries instead of attempting a separate, faulty recompilation. This respects the `Makefile` as the single source of truth for building.
- **tests/framework/test_kernel.h**:
    - Changed the type of `original_cr3` from `uint64_t` to `uint32_t` for 32-bit compatibility.
- **tests/framework/test_kernel.c**:
    - Replaced 64-bit specific inline assembly (`pushfq`) with its 32-bit equivalent (`pushfl`) in `test_kernel_save_state`.

**AI Assistant Stability Fix:**

- **userspace/ai_assistant.c**:
    - Fixed a critical buffer over-read bug in the keyword detection logic. The previous implementation read past the end of the input string, leading to undefined behavior and system freezes.
    - Replaced the unsafe logic with a robust implementation using `strstr` after converting the query to lowercase. This required adding `strstr` and `to_lowercase` helper functions to the file.
    - This change makes the AI assistant significantly more stable and correctly handles case-insensitive queries.

These changes make the project more robust and portable. The test suite should now pass once the required dependencies (`nasm`, `gcc-multilib`) are available in the build environment.

+++ tests/framework/test_kernel.h
@@ -27,7 +27,7 @@ typedef struct {
     test_ring_mode_t ring_mode;
     uint32_t kernel_stack;
     uint32_t user_stack;
-    uint64_t original_cr3;  // 64-bit pour compatibilité
+    uint32_t original_cr3;
     int interrupts_enabled;
     char test_heap[4096];
     size_t heap_used;
